refactor: use Handler in the StateHolder
