+++ shared/src/commonMain/kotlin/com/akjaw/ai/assistant/shared/dashboard/ui/Dashboard.kt
@@ -30,7 +30,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.akjaw.ai.assistant.shared.chat.domain.AddTaskFactory
+import com.akjaw.ai.assistant.shared.chat.domain.ChatMessageHandler
+import com.akjaw.ai.assistant.shared.chat.domain.createKtorClient
 import com.akjaw.ai.assistant.shared.chat.presentation.ChatScreenStateHolder
 import com.akjaw.ai.assistant.shared.chat.ui.ChatScreen
 import com.akjaw.ai.assistant.shared.dashboard.domain.ChatType
@@ -43,12 +44,12 @@ class DashboardScreenStateHolder(
     private val scope: CoroutineScope,
     private val pagerState: PagerState,
     initialType: ChatType = ChatType.Notion,
-    private val addTaskFactory: AddTaskFactory = AddTaskFactory(),
+    private val chatScreenStateHolder: ChatMessageHandler = ChatMessageHandler(createKtorClient()),
 ) {
 
     var stateHolders: List<ChatScreenStateHolder> by mutableStateOf(
         ChatType.values().map { type ->
-            ChatScreenStateHolder(type, scope, addTaskFactory.create(type))
+            ChatScreenStateHolder(type, scope, chatScreenStateHolder)
         }
     )
